adapter: split out a new timestamp-oracle crate
Besides the legacy impl of the trait (which is left in place) this is
nicely self-contained, so break it out into its own crate. The immediate
motivation is so I can depend on the trait (and maybe the postgres impl)
in persistcli without pulling in adapter.
